requirements.txt
coverage==7.13.5
iniconfig==2.3.0
packaging==26.1
pillow==10.4.0
pluggy==1.6.0
Pygments==2.20.0
pyperclip==1.9.0
pytest==9.0.3
pytest-cov==7.1.0
pyzbar==0.1.9
qrcode==8.0
ruff==0.15.10
